+++ editor/editor-core/example-helpers/kitchen-sink/kitchen-sink-adf-input.tsx
@@ -4,18 +4,20 @@ import { InputPadding, InputForm, Textarea } from './kitchen-sink-styles';
 
 export interface KitchenSinkAdfInputProps {
   value: string;
-  onSubmit(e: React.FormEvent): void;
-  onChange(e: React.ChangeEvent): void;
+  onSubmit?(e: React.FormEvent): void;
+  onChange?(e: React.ChangeEvent): void;
 }
 
 export const KitchenSinkAdfInput: React.StatelessComponent<KitchenSinkAdfInputProps> = props => {
   return (
     <InputForm>
       <InputPadding>
-        <Textarea value={props.value} onChange={props.onChange} />
-        <Button onClick={props.onSubmit} type="submit">
-          Import ADF
-        </Button>
+        <Textarea value={props.value} onChange={props.onChange} rows={20} />
+        {props.onSubmit && (
+          <Button onClick={props.onSubmit} type="submit">
+            Import ADF
+          </Button>
+        )}
       </InputPadding>
     </InputForm>
   );