+++ elements/user-picker/src/components/assets/microsoft.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+
+export function MicrosoftIcon(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      width="12"
+      height="12"
+      viewBox="0 0 12 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <g clipPath="url(#clip0)">
+        <path d="M0 0H5.70233V5.7023H0V0Z" fill="white" />
+        <path d="M6.29785 0H12.0002V5.7023H6.29785V0Z" fill="white" />
+        <path d="M0 6.29688H5.70233V11.9992H0V6.29688Z" fill="white" />
+        <path
+          d="M6.29785 6.29688H12.0002V11.9992H6.29785V6.29688Z"
+          fill="white"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0">
+          <rect width="12" height="11.9999" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}