+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-common
 
+## 52.0.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 52.0.0
 
 ### Major Changes