+++ elements/share/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/share
 
+## 0.24.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.24.1
 
 ### Patch Changes