+++ elements/user-picker/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/user-picker
 
+## 7.1.0
+
+### Minor Changes
+
+- [`c9327fc11e`](https://bitbucket.org/atlassian/atlassian-frontend/commits/c9327fc11e) - [ux] Add ability to user picker to display the external users along with the sources they come from.
+
+## 7.0.0
+
+### Major Changes
+
+- [`2c148c9eb9`](https://bitbucket.org/atlassian/atlassian-frontend/commits/2c148c9eb9) - Update the on Error prop to allow fail-over search client implementation for products.
+  Confluence fail-over client.
+
 ## 6.9.3
 
 ### Patch Changes